feat: Add utils module
